+++ client/src/main.tsx
@@ -3,7 +3,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 
 import App from './App.jsx';
 import Home from './pages/Home';
-import Checklist from './pages/Checklist';
+import Checklist from './pages/Signup.js';
 import Blog from './pages/Blog';
 import ErrorPage from './pages/Error';
